factor out InstructorTaskTestCase base class, for reuse in test_tasks.
